Moved sidecar test helpers back to package since it is used in vtexplain
