+++ go/vt/sidecardb/sidecardb.go
@@ -22,8 +22,11 @@ import (
 	"fmt"
 	"io/fs"
 	"path/filepath"
+	"regexp"
 	"runtime"
 	"strings"
+
+	"vitess.io/vitess/go/mysql/fakesqldb"
 
 	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/sqlparser"
@@ -360,3 +363,49 @@ func (si *schemaInit) loadExistingTables() error {
 	}
 	return nil
 }
+
+// region unit-test-only
+// this section uses helpers used in go/vt/vtexplain/vtexplain_vttablet.go, hence it is here and not in the _test.go file
+// query patterns to handle in mocks
+var sidecarDBInitQueryPatterns = []string{
+	ShowSidecarDatabasesQuery,
+	SelectCurrentDatabaseQuery,
+	CreateSidecarDatabaseQuery,
+	UseSidecarDatabaseQuery,
+	GetCurrentTablesQuery,
+	CreateTableRegexp,
+	AlterTableRegexp,
+}
+
+// AddSidecarDBSchemaInitQueries adds sidecar database schema related queries to a mock db
+func AddSidecarDBSchemaInitQueries(db *fakesqldb.DB) {
+	result := &sqltypes.Result{}
+	for _, q := range sidecarDBInitQueryPatterns {
+		db.AddQueryPattern(q, result)
+	}
+	for _, table := range sidecarTables {
+		result = sqltypes.MakeTestResult(sqltypes.MakeTestFields(
+			"Table|Create Table",
+			"varchar|varchar"),
+			fmt.Sprintf("%s|%s", table.name, table.schema),
+		)
+		db.AddQuery(fmt.Sprintf(ShowCreateTableQuery, table.name), result)
+	}
+}
+
+// MatchesSidecarDBInitQuery returns true if query has one of the test patterns as a substring, or it matches a provided regexp
+func MatchesSidecarDBInitQuery(query string) bool {
+	query = strings.ToLower(query)
+	for _, q := range sidecarDBInitQueryPatterns {
+		q = strings.ToLower(q)
+		if strings.Contains(query, q) {
+			return true
+		}
+		if match, _ := regexp.MatchString(q, query); match {
+			return true
+		}
+	}
+	return false
+}
+
+// endregion
