+++ go/vt/sidecardb/sidecardb_test.go
@@ -18,9 +18,6 @@ package sidecardb
 
 import (
 	"context"
-	"fmt"
-	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -116,46 +113,4 @@ func TestValidateSchema(t *testing.T) {
 			}
 		})
 	}
-}
-
-// query patterns to handle in mocks
-var sidecarDBInitQueryPatterns = []string{
-	ShowSidecarDatabasesQuery,
-	SelectCurrentDatabaseQuery,
-	CreateSidecarDatabaseQuery,
-	UseSidecarDatabaseQuery,
-	GetCurrentTablesQuery,
-	CreateTableRegexp,
-	AlterTableRegexp,
-}
-
-// AddSidecarDBSchemaInitQueries adds sidecar database schema related queries to a mock db
-func AddSidecarDBSchemaInitQueries(db *fakesqldb.DB) {
-	result := &sqltypes.Result{}
-	for _, q := range sidecarDBInitQueryPatterns {
-		db.AddQueryPattern(q, result)
-	}
-	for _, table := range sidecarTables {
-		result = sqltypes.MakeTestResult(sqltypes.MakeTestFields(
-			"Table|Create Table",
-			"varchar|varchar"),
-			fmt.Sprintf("%s|%s", table.name, table.schema),
-		)
-		db.AddQuery(fmt.Sprintf(ShowCreateTableQuery, table.name), result)
-	}
-}
-
-// MatchesSidecarDBInitQuery returns true if query has one of the test patterns as a substring, or it matches a provided regexp
-func MatchesSidecarDBInitQuery(query string) bool {
-	query = strings.ToLower(query)
-	for _, q := range sidecarDBInitQueryPatterns {
-		q = strings.ToLower(q)
-		if strings.Contains(query, q) {
-			return true
-		}
-		if match, _ := regexp.MatchString(q, query); match {
-			return true
-		}
-	}
-	return false
 }
